+++ asciimatics/renderers.py
@@ -361,13 +361,15 @@ class ColourImageFile(StaticRenderer):
     """
 
     def __init__(self, screen, filename, height=30, bg=Screen.COLOUR_BLACK,
-                 fill_background=False):
+                 fill_background=False, uni=False, dither=False):
         """
         :param screen: The screen to use when displaying the image.
         :param filename: The name of the file to render.
         :param height: The height of the text rendered image.
         :param bg: The default background colour for this image.
         :param fill_background: Whether to set background colours too.
+        :param uni: Whether to use unicode box characters or not.
+        :param dither: Whether to dither the rendered image or not.
         """
         super(ColourImageFile, self).__init__()
         with Image.open(filename) as image:
@@ -382,7 +384,8 @@ class ColourImageFile(StaticRenderer):
             for frame in _ImageSequence(image):
                 ascii_image = ""
                 frame = frame.resize(
-                    (int(frame.size[0] * height * 2.0 / frame.size[1]), height),
+                    (int(frame.size[0] * height * 2.0 / frame.size[1]),
+                     height * 2 if uni else height),
                     Image.BICUBIC)
                 tmp_img = Image.new("P", (1, 1))
                 tmp_img.putpalette(screen.palette)
@@ -393,7 +396,7 @@ class ColourImageFile(StaticRenderer):
                 tmp_img.load()
                 new_frame.load()
                 new_frame = new_frame._new(
-                    new_frame.im.convert("P", 0, tmp_img.im))
+                    new_frame.im.convert("P", 3 if dither else 0, tmp_img.im))
 
                 # Blank out any transparent sections of the image for complex
                 # images with alpha blending.
@@ -402,22 +405,36 @@ class ColourImageFile(StaticRenderer):
                         frame.split()[-1], lambda a: 255 if a <= 64 else 0)
                     new_frame.paste(16, mask)
 
+                # Decide what "brush" we're going to use for the rendering.
+                brush = "▄" if uni else "#"
+
                 # Convert the resulting image to coloured ASCII codes.
-                for py in range(0, new_frame.size[1]):
+                for py in range(0, new_frame.size[1], 2 if uni else 1):
+                    # Looks like some terminals need a character printed before
+                    # they really reset the colours - so insert a dummy char
+                    # to reset the background if needed.
+                    if uni:
+                        ascii_image += "${%d,2,%d}." % (bg, bg)
                     ascii_image += "\n"
                     for px in range(0, new_frame.size[0]):
                         real_col = frame.getpixel((px, py))
+                        real_col2 = (frame.getpixel((px, py + 1)) if uni else
+                            real_col)
                         col = new_frame.getpixel((px, py))
-                        if real_col == background or col == 16:
-                            if fill_background:
+                        col2 = new_frame.getpixel((px, py + 1)) if uni else col
+                        if ((real_col == real_col2 == background) or
+                                (col == col2 == 16)):
+                            if fill_background or uni:
                                 ascii_image += "${%d,2,%d} " % (bg, bg)
                             else:
                                 ascii_image += "${%d} " % bg
                         else:
-                            if fill_background:
-                                ascii_image += "${%d,2,%d}#" % (col, col)
+                            if fill_background or uni:
+                                ascii_image += "${%d,2,%d}%s" % (col2, col, brush)
                             else:
                                 ascii_image += "${%d}#" % col
+                if uni:
+                    ascii_image += "${%d,2,%d}." % (bg, bg)
                 self._images.append(ascii_image)